File who initializes a class Square that defines a square by: (based on 4-square.py)
Private instance attribute: size:
1) property def size(self): to retrieve it
2) property setter def size(self, value): to set it:
a) size must be an integer, otherwise raise a TypeError exception with the message size must be an integer
b) if size is less than 0, raise a ValueError exception with the message size must be >= 0
Instantiation with optional size: def __init__(self, size=0):
3) Public instance method: def area(self): that returns the current square area
4) Public instance method: def my_print(self): that prints in stdout the square with the character #:
a) if size is equal to 0, print an empty line.
b) print (new line * position[1]) before printing the square.
c) print (space * position[0] before printing a line of "#"